requirements.txt
streamlit>=1.28.0
neo4j>=5.14.0
plotly>=5.17.0
requests>=2.31.0
beautifulsoup4>=4.12.2
openai>=1.3.0
python-dotenv>=1.0.0
pandas>=2.0.0
tiktoken>=0.5.1
streamlit-agraph>=0.0.45
watchdog>=6.0.0